requirements.txt
obs-websocket-py==0.5.3
six==1.15.0
websocket-client==0.58.0